rename materials folder
