+++ src/app/shared/models/render/cavern-seer-scan.render-model.ts
@@ -3,7 +3,7 @@ import { BufferGeometry, Group, Line, LineBasicMaterial, Mesh, MeshBasicMaterial
 import { float4x4ToMatrix4 } from "../../functions/float4x4-to-matrix4";
 import { markSceneOfItemForReRender } from "../../functions/mark-scene-of-item-for-rerender";
 import { ignoreNullishArray } from "../../operators/ignore-nullish";
-import { BaseMaterialService } from "../../services/3d-managers/base-material.service";
+import { BaseMaterialService } from "../../services/materials/base-material.service";
 import type { CSMeshSnapshot, SurveyLine, SurveyStation } from "../../types/cavern-seer-scan";
 import { BaseAnnotation } from "../annotations/base.annotation";
 import { TemporaryAnnotation } from "../annotations/temporary.annotation";
