+++ src/app/shared/services/materials/mesh-standard-material.service.ts
